requirements.txt
coverage==7.2.2
Django==4.1.7
django-admin-thumbnails==0.2.6
djangorestframework==3.14.0
django-cors-headers==3.14.0
djoser==2.1.0
Pillow==9.4.0
psycopg2==2.9.5
stripe==5.3.0
uWSGI==2.0.21
